cloudsoc-x/requirements.txt
# Core
boto3==1.34.0
requests==2.31.0
requests-aws4auth==1.3.1
python-dotenv==1.0.1

# API server
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0

# Dev / testing
pytest==8.1.1
pytest-cov==5.0.0
pytest-mock==3.14.0
flake8==7.0.0
